requirements.txt
google-genai==1.0.0
python-gitlab==4.4.0
python-dotenv==1.0.1
pathspec==0.12.1
gitpython==3.1.43
pydantic==2.10.0
pytest==8.3.0

